# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: RDV.BI API
  version: 0.1.0
  summary: Appointment booking, queue, and institution administration API
  description: |
    OpenAPI contract for the current RDV.BI backend.

    Notes:
    - Public endpoints live under `/api/public`.
    - Authentication endpoints live under `/api/auth`.
    - Admin endpoints require a bearer JWT issued by `/api/auth/login`.
    - `POST /api/auth/bootstrap-admin` exists only for bootstrap and should be removed or disabled in production.
servers:
  - url: http://localhost:4000
    description: Local development
...
paths:
...
  /api/admin/institutions/pending:
    get:
      tags: [Admin Institutions]
      summary: List pending institutions
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Pending institutions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Institution'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/admin/institutions/{id}/approve:
    post:
      tags: [Admin Institutions]
      summary: Approve an institution
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UuidIdParam'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApprovalRequest'
      responses:
        '200':
          description: Approved institution
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Institution'
        '404':
          $ref: '#/components/responses/ErrorResponse'
  /api/admin/institutions/{id}/reject:
    post:
      tags: [Admin Institutions]
      summary: Reject an institution
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UuidIdParam'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApprovalRequest'
      responses:
        '200':
          description: Rejected institution
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Institution'
        '404':
          $ref: '#/components/responses/ErrorResponse'
  /api/admin/institutions/{id}:
    put:
      tags: [Admin Institutions]
      summary: Update institution details
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UuidIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InstitutionUpdateRequest'
      responses:
        '200':
          description: Updated institution
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Institution'
        '404':
          $ref: '#/components/responses/ErrorResponse'
  /api/admin/institutions/{id}/services:
    post:
      tags: [Admin Services]
      summary: Create a service for an institution
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UuidIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceRequest'
      responses:
        '201':
          description: Created service
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Service'
        '404':
          $ref: '#/components/responses/ErrorResponse'
    get:
      tags: [Admin Services]
      summary: List all services for an institution
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UuidIdParam'
      responses:
        '200':
          description: Services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Service'
        '404':
          $ref: '#/components/responses/ErrorResponse'
  /api/admin/services/{serviceId}:
    put:
      tags: [Admin Services]
      summary: Update a service
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ServiceIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceUpdateRequest'
      responses:
        '200':
          description: Updated service
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Service'
        '404':
          $ref: '#/components/responses/ErrorResponse'
    delete:
      tags: [Admin Services]
      summary: Delete a service
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ServiceIdParam'
      responses:
        '204':
          description: Deleted
        '404':
          $ref: '#/components/responses/ErrorResponse'
  /api/admin/institutions/{id}/time-slots:
    get:
      tags: [Admin Time Slots]
      summary: List institution time slots
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UuidIdParam'
        - name: date
          in: query
          required: false
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Time slots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TimeSlot'
        '404':
          $ref: '#/components/responses/ErrorResponse'
    post:
      tags: [Admin Time Slots]
      summary: Create one or more time slots
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UuidIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/TimeSlotRequest'
      responses:
        '201':
          description: Created time slots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TimeSlot'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
  /api/admin/time-slots/{slotId}:
    put:
      tags: [Admin Time Slots]
      summary: Update a time slot
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SlotIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TimeSlotUpdateRequest'
      responses:
        '200':
          description: Updated time slot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimeSlot'
        '404':
          $ref: '#/components/responses/ErrorResponse'
    delete:
      tags: [Admin Time Slots]
      summary: Delete a time slot
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SlotIdParam'
      responses:
        '204':
          description: Deleted
        '404':
          $ref: '#/components/responses/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    UuidIdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    ServiceIdParam:
      name: serviceId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    SlotIdParam:
      name: slotId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    ErrorResponse:
      description: Error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Authenticated but not allowed to access the resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
      example:
        error: Institution not found
...
    InstitutionUpdateRequest:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        type:
          type: string
        address:
          type: string
        phone:
          type: string
        email:
          type: string
          format: email
        logo_url:
          type: string
        is_active:
          type: boolean
    ApprovalRequest:
      type: object
      properties:
        notes:
          type: string
    ServiceRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        durationMinutes:
          type: integer
          minimum: 1
        prefix:
          type: string
        active:
          type: boolean
      required: [name, durationMinutes]
    ServiceUpdateRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        duration_minutes:
          type: integer
        prefix:
          type: string
        is_active:
          type: boolean
    TimeSlotRequest:
      type: object
      properties:
        date:
          type: string
          format: date
        startTime:
          type: string
          format: time
        endTime:
          type: string
          format: time
        capacity:
          type: integer
          minimum: 1
      required: [date, startTime, endTime, capacity]
    TimeSlotUpdateRequest:
      type: object
      properties:
        date:
          type: string
          format: date
        start_time:
          type: string
          format: time
        end_time:
          type: string
          format: time
        capacity:
          type: integer
        is_blocked:
          type: boolean
        block_reason:
          type: string
...
    Institution:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        type:
          type: string
        contactEmail:
          type: string
          format: email
        contactPhone:
          type: string
        address:
          type: string
        description:
          type: string
        logoUrl:
          type: string
        locationLat:
          type: number
        locationLng:
          type: number
        status:
          $ref: '#/components/schemas/InstitutionStatus'
        active:
          type: boolean
        approvalNotes:
          type: string
        approvedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, slug, status, active]
    Service:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        durationMinutes:
          type: integer
        prefix:
          type: string
        active:
          type: boolean
      required: [id, name, durationMinutes, active]
    TimeSlot:
      type: object
      properties:
        id:
          type: string
          format: uuid
        date:
          type: string
          format: date
        startTime:
          type: string
          format: time
        endTime:
          type: string
          format: time
        capacity:
          type: integer
        bookedCount:
          type: integer
        blocked:
          type: boolean
        blockReason:
          type: string
      required: [id, date, startTime, endTime, capacity, bookedCount, blocked]
...
    InstitutionStatus:
      type: string
      enum: [PENDING, APPROVED, REJECTED]
...